Refactor product and user entities, add product repository and service
- Deleted Order.java entity
- Modified Product.java and User.java entities
- Added ProductRepository and ProductService
- Updated application.properties

+++ src/main/java/com/example/demo/entity/Product.java
@@ -2,12 +2,16 @@ package com.example.demo.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+
+import org.springframework.data.annotation.Id;
 
 @Entity
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer Id;
+    private Long Id;
 
     @Column(nullable = false)
     private String name;
@@ -16,7 +20,7 @@ public class Product {
 
     @Column(nullable = false)
     private String category;
-    
+
     @Column(nullable = false)
     private String imageUrl;
